Reorganise directory structure and include path in tsconfig.base.json
