+++ apps/wallet-connect/src/services/RemixClient.js
