+++ apps/wallet-connect/src/services/serviceWorker.js
