+++ tsconfig.paths.json
@@ -151,6 +151,9 @@
             "@remix-project/ghaction-helper": [
                 "libs/ghaction-helper/src/index.ts"
             ],
+            "@remixproject/wallet-connect-plugin": [
+                "apps/wallet-connect/src/index.ts"
+            ],
         }
     }
 }
